Optimisation des pièces et du loot.
